add/refactor my code into master with master tims help.
