--- src/main/java/parser/HeaderInfo.java
@@ -1,7 +1,0 @@
-package parser;
-
-public class HeaderInfo {
-    public String Title;
-    public String Author;
-
-}
